refactor(orchestrator): split pipeline.rs analyzer logic into analyze.rs
Pre-emptive Rule 4 compliance step before #188 (analyzer wiring) lands.
Extracts analyze_di_graph from pipeline.rs (was 311 LOC) into a new
pipeline/analyze.rs submodule. Zero behavior change — function body
moved byte-for-byte; only delta is the cosmetic std::fs::read_to_string
shortened to fs::read_to_string via a use statement.

Validated by tyrus-architect + everything-claude-code:code-reviewer
(approved). All gates pass: cargo fmt, clippy --workspace --all-targets
-D warnings, nextest 259/259 + 1 pre-existing skip.

After split: pipeline.rs 287 LOC, analyze.rs 32 LOC. Headroom for #188
to add ~25-50 LOC of error aggregation in analyze.rs without pushing
pipeline.rs over 400.

Closes #195
Refs: .claude/plan/uat-bug-fixes-2026-05-07.md (D7, PR-0)
